+++ src/cli/churros-add.js
@@ -4,6 +4,7 @@ const commander = require('commander');
 const fs = require('fs');
 const inquirer = require('inquirer');
 const replaceStream = require('replacestream');
+const tools = require('core/tools');
 
 const terminate = (msg, args) => {
   args ? console.error(msg, args) : console.error(msg);
@@ -26,6 +27,13 @@ const buildQuestion = (name, type, message, validate, isDefault) => new Object({
   default: isDefault
 });
 
+const buildChoiceQuestion = (name, type, message, choices, filter) => {
+  const q = buildQuestion(name, type, message);
+  q.choices = choices;
+  q.filter = filter;
+  return q;
+};
+
 const buildPlatformQuestions = () =>
   new Promise((res, rej) => {
     res([buildQuestion('name', 'input', 'Platform resource name:', (value) => validateValue(value))]);
@@ -36,6 +44,7 @@ const buildElementQuestions = () =>
     res([
       buildQuestion('name', 'input', 'Element name:', (value) => validateValue(value)),
       buildQuestion('hub', 'input', 'Hub name:', (value) => validateValue(value)),
+      buildChoiceQuestion('auth', 'list', 'Auth type:', ['standard', 'oauth1', 'oauth2'], (value) => value.toLowerCase())
     ]);
   });
 
@@ -54,6 +63,66 @@ const askResourceQuestions = (r, resources) => {
       if (answers.more) askResourceQuestions(r, resources).then(resource => res(resource));
       else {
         r.resources = resources;
+        res(r);
+      }
+    });
+  });
+};
+
+const buildConfigQuestions = (auth, currentProps, firstTime) => {
+  const qs = [];
+  // these props are ALWAYS required for OAuth2 and OAuth1 elements
+  if (firstTime && (auth === 'oauth1' || auth === 'oauth2')) {
+    qs.push(buildQuestion('oauth.api.key', 'input', 'OAuth API Key:', (value) => validateValue(value)));
+    qs.push(buildQuestion('oauth.api.secret', 'input', 'OAuth API Secret:', (value) => validateValue(value)));
+    qs.push(buildQuestion('username', 'input', 'Username:', (value) => validateValue(value)));
+    qs.push(buildQuestion('password', 'input', 'Password:', (value) => validateValue(value)));
+    qs.push(buildQuestion('oauth.scope', 'input', 'OAuth Scope (optional):'));
+    qs.push(buildQuestion('site.address', 'input', 'Subdomain (optional):'));
+    qs.push(buildQuestion('more', 'confirm', 'Add more configs:'));
+  } else if (firstTime) {
+    qs.push(buildQuestion('more', 'confirm', 'Add configs:'));
+  }
+
+  if (!firstTime) {
+    // using a random string so we can matchup these from key: "", value: "" to key: value later on
+    const random = tools.random();
+    qs.push(buildQuestion('key.' + random, 'input', 'Key (i.e. my.config.key):'));
+    qs.push(buildQuestion('value.' + random, 'input', 'Value:'));
+    qs.push(buildQuestion('more', 'confirm', 'Add more configs:'));
+  }
+
+  return qs;
+};
+
+const askConfigQuestions = (r, config) => {
+  config = config || {};
+  return new Promise((res, rej) => {
+    const firstTime = Object.keys(config).length <= 0;
+    inquirer.prompt(buildConfigQuestions(r.auth, config, firstTime), (answers) => {
+      Object.keys(answers).forEach(key => config[key] = answers[key]);
+      if (answers.more) askConfigQuestions(r, config).then(innerR => res(innerR));
+      else {
+        delete config.more; // cleanup the 'more' property
+
+        // combine the key/values
+        Object.keys(config)
+          .filter(c => c.startsWith('key.'))
+          .reduce((p, c) => {
+            const rando = c.split('.')[1]; // key.aeiond will ouput aeiond
+            const key = 'key.' + rando;
+            const value = 'value.' + rando;
+            config[config[key]] = config[value];
+            delete config[key]; // remove the key.randomstring
+            delete config[value]; // remove the value.randomstring
+          }, {});
+
+        // delete any configs that have empty values
+        Object.keys(config)
+          .filter(c => !config[c])
+          .reduce((p, c) => delete config[c], {});
+
+        r.config = config;
         res(r);
       }
     });
@@ -84,10 +153,9 @@ const stubPlatformFiles = (r) =>
     res(r);
   });
 
-const stubElementFiles = (r) =>
+const stubElementFiles = (answers) =>
   new Promise((res, rej) => {
-    const rootDir = buildRootElementDir();
-    const suiteDir = rootDir + '/' + r.name;
+    const suiteDir = buildRootElementDir() + '/' + answers.name;
     if (!fs.existsSync(suiteDir)) fs.mkdirSync(suiteDir);
 
     const assetsDir = suiteDir + '/assets';
@@ -96,7 +164,19 @@ const stubElementFiles = (r) =>
     const transformationsFile = assetsDir + '/transformations.json';
     if (!fs.existsSync(transformationsFile)) fs.writeFileSync(transformationsFile, '{}');
 
-    r.resources.forEach(resource => {
+    const config = answers.config;
+    if (config) {
+      let saucePath = process.env.HOME + '/.churros/sauce.json';
+      let sauceConfig = require(saucePath);
+      Object.keys(config).forEach(c => {
+        const value = config[c];
+        if (!sauceConfig[answers.name]) sauceConfig[answers.name] = {}; // init
+        sauceConfig[answers.name][c] = value;
+      });
+      fs.writeFileSync(saucePath, JSON.stringify(sauceConfig, null, 2));
+    }
+
+    answers.resources.forEach(resource => {
       const name = resource.name;
       const vendorName = resource.vendorName;
       const vendorId = resource.vendorId;
@@ -106,7 +186,7 @@ const stubElementFiles = (r) =>
       if (!fs.existsSync(resourceJsFile)) {
         fs.createReadStream(__dirname + '/assets/element.suite.template.js')
           .pipe(replaceStream('%resource', name))
-          .pipe(replaceStream('%hub', r.hub))
+          .pipe(replaceStream('%hub', answers.hub))
           .pipe(replaceStream('%user', process.env.USER))
           .pipe(fs.createWriteStream(resourceJsFile));
       } else {
@@ -134,11 +214,11 @@ const stubElementFiles = (r) =>
         };
         fs.writeFileSync(transformationsFile, JSON.stringify(currentTransformations, null, 2));
       } else {
-        console.log('transformation for %s already exists for %s', name, r.name);
+        console.log('transformation for %s already exists for %s', name, answers.name);
       }
     });
 
-    res(r);
+    res(answers);
   });
 
 const complete = (r, link) => {
@@ -146,7 +226,7 @@ const complete = (r, link) => {
   console.log('Successfully stubbed out test suite for %s', r.name);
   console.log('');
   console.log('For information on next steps:');
-  console.log('      https://github.com/cloud-elements/churros/blob/master/CONTRIBUTING.md#%s', link);
+  console.log('    https://github.com/cloud-elements/churros/blob/master/CONTRIBUTING.md#%s', link);
   console.log('');
 };
 
@@ -154,6 +234,7 @@ const add = (type, options) => {
   if (type === 'element') {
     buildElementQuestions(type)
       .then(r => askQuestions(r))
+      .then(r => askConfigQuestions(r))
       .then(r => askResourceQuestions(r))
       .then(r => stubElementFiles(r))
       .then(r => complete(r, 'new-element-suite'))
@@ -170,10 +251,13 @@ const add = (type, options) => {
 };
 
 commander
-  .command('suite type', '[element || platform]')
+  .command('type', '[element || platform]')
   .action((type, options) => add(type, options))
   .on('--help', () => {
     console.log('  Examples:');
+    console.log('');
+    console.log('    $ churros add element');
+    console.log('    $ churros add platform');
     console.log('');
   })
   .parse(process.argv);