+++ src/core/props.js
@@ -1,20 +1,20 @@
 'use strict';
 
-const fs = require('fs');
 const tools = require('core/tools');
 const util = require('util');
+const logger = require('winston');
 
-const missingPropForRootKey = (rootKey, key) => {
-  const msg = util.format("Missing required property '%s' for '%s'\n   Note: Can set this property by calling 'churros props %s:%s <value>'", key, rootKey, rootKey, key);
-  throw new Error(msg);
-};
+let config = {};
 
-var exports = module.exports = (config) => {
-  const configFile = process.env.HOME + '/.churros/sauce.json';
+var exports = module.exports = (initConfig) => {
+  // if no config was passed in, then stop
+  if (!initConfig) tools.logAndThrow('Cannot initialize props with empty config.');
+  config = initConfig;
 
-  // if no config was passed in AND the sauce.json file does not exist, then stop
-  if (!config && !fs.existsSync(configFile)) tools.logAndThrow('No props file found.  Please run churros init to get started.');
-  if (!config) config = require(configFile);
+  const missingPropForRootKey = (rootKey, key) => {
+    const msg = util.format("Missing required property '%s' for '%s'\n   Note: Can set this property by calling 'churros props %s:%s <value>'", key, rootKey, rootKey, key);
+    throw Error(msg);
+  };
 
   exports.get = (key) => {
     var value = config[key];
@@ -23,7 +23,7 @@ var exports = module.exports = (config) => {
     // right now, if a property isn't found we just fail immediately.  long term, we want to add the ability to run in
     // '--prompt' mode, which will prompt the user for a value to use.
     const msg = util.format("No value found for required property '%s'\n   Note: Can set this property by calling 'churros props %s <value>'", key, key);
-    throw new Error(msg);
+    throw Error(msg);
   };
 
   exports.getForKey = (rootKey, key) => {
@@ -36,6 +36,7 @@ var exports = module.exports = (config) => {
   };
 
   exports.getOptionalForKey = (rootKey, key) => {
+    logger.debug('Looking for optional property %s:%s', rootKey, key);
     const elementProps = config[rootKey];
     if (!elementProps) return null;
 
@@ -55,9 +56,10 @@ var exports = module.exports = (config) => {
   };
 
   exports.all = (element) => {
+    logger.debug('Looking for props for %s', element);
     if (!config[element]) {
       const msg = util.format("No properties found for element '%s'\n   Note: Can setup properties for this element by calling 'churros props %s:my.config.key <value>'", element, element);
-      throw new Error(msg);
+      throw Error(msg);
     }
 
     const response = {};