+++ src/test/lifecycle.js
@@ -1,42 +1,53 @@
 'use strict';
 
+require('core/assertions');
 const chakram = require('chakram');
-const util = require('util');
-const argv = require('optimist').demand('user').argv;
-const props = require('core/props')();
-const tools = require('core/tools');
+const expect = chakram.expect;
+const argv = require('optimist').argv;
+const defaults = require('core/defaults');
+const logger = require('core/logger')(argv.verbose ? 'silly' : 'info');
 
-const setGlobalProps = (args) => {
-  props.set('user', args.user);
-  props.set('password', args.password);
-  props.set('url', args.url);
-  props.setForKey('events', 'wait', args.wait);
-  props.setForKey('events', 'load', args.load);
-  props.setForKey('events', 'element', args.loadElement);
-};
+let config;
+try {
+  config = require(process.env.HOME + '/.churros/sauce.json');
+} catch (e) {
+  console.log('No properties found.  Make sure to run \'churros init\' first.');
+  process.exit(1);
+}
+
+config.user = (argv.user || config.user);
+config.password = (argv.password || config.password);
+config.url = (argv.url || config.url);
+config.browser = (argv.browser || 'phantomjs');
+
+if (!config.events) config.events = {};
+config.events.wait = (argv.wait || config.events.wait);
+config.events.load = (argv.load || config.events.load);
+config.events.element = (argv.loadElement || config.events.element);
+
+// this happens once
+const props = require('core/props')(config);
 
 before((done) => {
-  tools.addCustomAssertions();
-  setGlobalProps(argv);
+  const url = props.get('url');
+  const secUrl = url + '/elements/j_spring_security_check';
+  const options = { jar: true, form: { j_username: props.get('user'), j_password: props.get('password') } };
 
-  const secUrl = props.get('url') + '/elements/j_spring_security_check';
-  const form = { j_username: props.get('user'), j_password: props.get('password') };
-  const options = { jar: true, form: form };
-  const secretOptions = { jar: true };
-
-  chakram.post(secUrl, null, options)
-    .then(r => chakram.get(props.get('url') + '/elements/api-v1/ui/getSecrets', secretOptions))
+  return chakram.post(secUrl, null, options)
     .then(r => {
-      props.set('user.secret', r.body.user);
-      props.set('org.secret', r.body.company);
-      chakram.setRequestDefaults({
-        baseUrl: props.get('url') + '/elements/api-v2',
-        headers: { Authorization: util.format('User %s, Organization %s', r.body.user, r.body.company) }
-      });
+      const secretOptions = { jar: true };
+      expect(r).to.have.statusCode(200);
+      return chakram.get(url + '/elements/api-v1/ui/getSecrets', secretOptions);
+    })
+    .then(r => {
+      expect(r).to.have.statusCode(200);
+      defaults(url + '/elements/api-v2', r.body.user, r.body.company);
       done();
     })
     .catch(r => {
-      console.log('Well shucks...failed to finish initialization...\n  Is %s up and running?\n  Do you have the right username and password?', props.get('url'));
+      // if the lifecycle fails, then we want to exit with an error and not let anything else continue
+      logger.error('Well shucks...failed to finish initialization...\n  Is %s up and running?\n  Do you have the right username and password?\n', url);
+      logger.error(r);
       process.exit(1);
     });
 });