+++ test/core/tester.test.js
@@ -1,30 +1,24 @@
 'use strict';
 
+require('core/assertions');
 const tester = require('core/tester');
 const nock = require('nock');
 const chakram = require('chakram');
 const expect = chakram.expect;
-const tools = require('core/tools');
 const fs = require('fs');
 
 const baseUrl = 'https://api.cloud-elements.com/elements/api-v2;';
 const auth = 'User fake, Organization fake';
 const eiId = 789;
 
-const headers = () => {
-  return {
-    reqheaders: { 'Authorization': (value) => value === auth }
-  };
-};
+const headers = () => new Object({ reqheaders: { 'Authorization': (value) => value === auth } });
 
-const eventHeaders = () => {
-  return {
-    reqheaders: {
-      'Authorization': (value) => value === auth,
-      'Element-Instances': (value) => value === eiId
-    }
-  };
-};
+const eventHeaders = () => new Object({
+  reqheaders: {
+    'Authorization': (value) => value === auth,
+    'Element-Instances': (value) => value === eiId
+  }
+});
 
 const genPayload = (opts) => {
   opts = opts || {};
@@ -43,14 +37,14 @@ const genSchema = () => new Object({
   required: ['id', 'foo']
 });
 
-const setup = () => {
-  tools.addCustomAssertions();
+beforeEach(() => {
   chakram.setRequestDefaults({
     baseUrl: baseUrl,
     headers: { Authorization: auth }
   });
 
   /** MOCKING OUT HTTP ENDPOINTS **/
+  /** https://github.com/pgte/nock#specifying-hostname **/
 
   /** POST **/
   nock(baseUrl, headers())
@@ -79,7 +73,10 @@ const setup = () => {
       return { message: 'No foo found with the given ID' };
     })
     .get('/foo')
-    .reply(200, (uri, requestBody) => [genPayload({ id: 123 }), genPayload({ id: 456 })]);
+    .reply(200, (uri, requestBody) => [genPayload({ id: 123 }), genPayload({ id: 456 })])
+    .get('/foo')
+    .query({ page: '1', pageSize: '1' })
+    .reply(200, () => genPayload({ id: 123 }));
 
   /** PATCH && PUT **/
   nock(baseUrl, headers())
@@ -88,10 +85,18 @@ const setup = () => {
       requestBody.id = 123;
       return requestBody;
     })
+    .patch('/foo/456')
+    .reply(404, (uri, requestBody) => {
+      return { message: 'No foo found with the given ID' };
+    })
     .put('/foo/123')
     .reply(200, (uri, requestBody) => {
       requestBody.id = 123;
       return requestBody;
+    })
+    .put('/foo/456')
+    .reply(404, (uri, requestBody) => {
+      return { message: 'No foo found with the given ID' };
     });
 
   /** DELETE **/
@@ -100,57 +105,46 @@ const setup = () => {
     .reply(200, (uri, requestBody) => {
       return {};
     });
-};
+});
 
 describe('tester', () => {
-  it('should support post', () => {
-    setup();
-    return tester.post('/foo', genPayload(), genSchema());
-  });
+  it('should support post', () => tester.post('/foo', genPayload(), genSchema()));
+  it('should support post with custom validation', () => tester.post('/foo/bad', {}, (r) => expect(r).to.have.statusCode(400)));
+  it('should support post with no schema or custom validation', () => tester.post('/foo', genPayload()));
 
-  it('should support post with custom validation', () => {
-    setup();
-    return tester.post('/foo/bad', {}, (r) => expect(r).to.have.statusCode(400));
-  });
+  it('should support get', () => tester.get('/foo/123', genSchema()));
+  it('should support get with custom validation', () => tester.get('/foo/456', (r) => expect(r).to.have.statusCode(404)));
+  it('should support get with no schema or custom validation', () => tester.get('/foo/123'));
 
-  it('should support post with no schema or custom validation', () => {
-    setup();
-    return tester.post('/foo', genPayload());
-  });
+  it('should support put', () => tester.put('/foo/123', genPayload(), genSchema()));
+  it('should support put with custom validation', () => tester.put('/foo/456', genPayload(), (r) => expect(r).to.have.statusCode(404)));
+  it('should support put with no schema or custom validation', () => tester.put('/foo/123', genPayload()));
 
-  it('should support get', () => {
-    setup();
-    return tester.get('/foo/123', genSchema());
-  });
+  it('should support patch', () => tester.patch('/foo/123', genPayload({ id: 123 }), genSchema()));
+  it('should support patch with custom validation', () => tester.patch('/foo/456', genPayload(), (r) => expect(r).to.have.statusCode(404)));
+  it('should support patch with no schema or custom validation', () => tester.put('/foo/123', genPayload()));
 
-  it('should support get with custom validation', () => {
-    setup();
-    return tester.get('/foo/456', (r) => expect(r).to.have.statusCode(404));
-  });
+  it('should support delete', () => tester.delete('/foo/123'));
 
-  it('should support put', () => {
-    setup();
-    return tester.put('/foo/123', genPayload(), genSchema());
-  });
+  it('should support find', () => tester.find('/foo', genSchema()));
 
-  it('should support patch', () => {
-    setup();
-    return tester.patch('/foo/123', genPayload({ id: 123 }), genSchema());
-  });
-
-  it('should support delete', () => {
-    setup();
-    return tester.delete('/foo/123');
-  });
-
-  it('should support find', () => {
-    setup();
-    return tester.find('/foo', genSchema());
+  it('should support crd', () => tester.crd('/foo', genPayload(), genSchema()));
+  it('should support crds', () => tester.crds('/foo', genPayload(), genSchema()));
+  it('should support crud', () => tester.crud('/foo', genPayload(), genSchema()));
+  it('should support cruds', () => tester.cruds('/foo', genPayload(), genSchema()));
+  it('should support creating events', () => tester.createEvents('myelement', eiId, genPayload(), 2));
+  it('should support listening for events', (done) => {
+    const port = 8085;
+    tester.listenForEvents(port, 1, 5)
+      .then(r => done())
+      .then(r => {
+        throw Error('Failed...');
+      });
+    chakram.setRequestDefaults({ headers: { 'User-Agent': 'churros-test' } });
+    return chakram.post('http://localhost:' + port, { event: 'green hat' });
   });
 
   it('should support post file', () => {
-    setup();
-
     // should really NOT depend on the file system here :/
     const filePath = '.tmp';
     fs.closeSync(fs.openSync(filePath, 'w'));
@@ -160,28 +154,22 @@ describe('tester', () => {
     return true;
   });
 
-  it('should support crd', () => {
-    setup();
-    return tester.crd('/foo', genPayload(), genSchema());
-  });
+  tester.test.badPost400('/foo/bad', genPayload());
+  tester.test.badPatch404('/foo', genPayload(), 456);
+  tester.test.badGet404('/foo', 456);
+  tester.test.cruds('/foo', genPayload(), genSchema());
+  tester.test.cruds('/foo', genPayload(), genSchema(), chakram.put);
+  tester.test.crud('/foo', genPayload(), genSchema());
+  tester.test.crd('/foo', genPayload(), genSchema());
+  tester.test.crds('/foo', genPayload(), genSchema());
+  tester.test.create('/foo', genPayload(), genSchema());
+  tester.test.paginate('/foo', genSchema(), {});
 
-  it('should support crds', () => {
-    setup();
-    return tester.crds('/foo', genPayload(), genSchema());
+  tester.for('fakehub', 'resource');
+  tester.for('fakehub', 'resource', (api) => {
+    it('should support the for with a hub passed in', () => expect(api).to.equal('/hubs/fakehub/resource'));
   });
-
-  it('should support crud', () => {
-    setup();
-    return tester.crud('/foo', genPayload(), genSchema());
-  });
-
-  it('should support cruds', () => {
-    setup();
-    return tester.cruds('/foo', genPayload(), genSchema());
-  });
-
-  it('should support creating events', () => {
-    setup();
-    return tester.createEvents('myelement', eiId, genPayload(), 2);
+  tester.for(null, 'platformresource', (api) => {
+    it('should support the for with a hub passed in', () => expect(api).to.equal('/platformresource'));
   });
 });